libhwsec: Pick appropriate ninja target
Otherwise, I'm not sure we're guaranteed what ninja will pick. Or at
least, I think it picks up more targets (that we don't want; like
trousers, even when we're not building for tpm1.2) if we omit this.

This is covered by platform.eclass and platform2/common-mk/platform2.py,
but we've elided that so far.

Signed-off-by: Brian Norris <briannorris@chromium.org>